Fix CI: Implement 2-pass build for Wine bridge
- Resolves package conflict between libwine-dev:i386 and libwine-dev:amd64 on Ubuntu 22.04.
- Pass 1: Installs 32-bit tools, builds NPClient32/FreeTrackClient32, stashes libs.
- Pass 2: Installs 64-bit tools, builds remaining targets.
- Inject 32-bit libs into AppImage during bundling.

+++ src/wine_bridge/client/NPClient_main.c
@@ -45,7 +45,7 @@ static void dbg_report(const char *msg, ...) {
 }
 
 BOOL WINAPI DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved) {
-  TRACE("(0x%p, %lu, %p)\n", hinstDLL, fdwReason, lpvReserved);
+  TRACE("(0x%p, %u, %p)\n", hinstDLL, (unsigned int)fdwReason, lpvReserved);
   thisDll = hinstDLL;
   switch (fdwReason) {
 #ifdef DLL_WINE_PREATTACH
